implement oversampling

+++ Source/Biquad.h
@@ -58,8 +58,8 @@ struct Biquad {
 
     inline sample_t processSample(sample_t sample) {
 
-        sample_t outputSample = sample*b0 + x1*b1 + x2*b2 
-                              - y1*a1 - y2*a2;
+        sample_t outputSample = (float)(sample * b0 + x1 * b1 + x2 * b2
+                              - y1 * a1 - y2 * a2);
         x2 = x1;
         x1 = sample;
         y2 = y1;
